moj_prvy_backend/requirements.txt
Flask
gunicorn
groq
python-dotenv
psycopg2-binary
